+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
---
+- Periodogram(freqs: ArrayLike | None = None) is added to set fixed user-defined frequency grids
 
 ### Changed